requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
designer==0.4.2
Django==5.0.3
django-ckeditor==6.7.0
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-filter==24.2
django-js-asset==2.2.0
django-phonenumber-field==8.0.0
django-resized==1.0.2
djangorestframework==3.15.1
idna==3.6
joblib==1.3.2
nltk==3.8.1
phonenumberslite==8.13.40
pillow==10.2.0
pygame==2.5.2
regex==2023.12.25
requests==2.31.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.0
whitenoise==6.6.0
